+++ server/src/main/java/org/opensearch/extensions/ExtensionsManager.java
@@ -365,6 +365,8 @@ public class ExtensionsManager {
             public void handleResponse(InitializeExtensionResponse response) {
                 for (DiscoveryExtensionNode extension : extensionIdMap.values()) {
                     if (extension.getName().equals(response.getName())) {
+                        // TODO Generate secret for extension
+
                         extensions.add(extension);
                         logger.info("Initialized extension: " + extension.getName());
                         break;
@@ -384,6 +386,7 @@ public class ExtensionsManager {
                 return ThreadPool.Names.GENERIC;
             }
         };
+
         try {
             logger.info("Sending extension request type: " + REQUEST_EXTENSION_ACTION_NAME);
             transportService.connectToExtensionNode(extension);
@@ -485,6 +488,7 @@ public class ExtensionsManager {
             public void handleResponse(IndicesModuleResponse response) {
                 logger.info("received {}", response);
                 if (response.getIndexEventListener() == true) {
+                    logger.info("indexModule.addIndexEventListener");
                     indexModule.addIndexEventListener(new IndexEventListener() {
                         @Override
                         public void beforeIndexRemoved(
@@ -514,7 +518,9 @@ public class ExtensionsManager {
                             }
                         }
                     });
+                    inProgressFuture.complete(response);
                 } else {
+                    logger.info("Probably shouldn't be here");
                     inProgressFuture.complete(response);
                 }
             }